+++ src/main/java/com/stock/market/dto/StockRequestAndResponse.java
@@ -4,4 +4,4 @@ import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
-public record StockRequestAndResponse(@NotBlank String name, @NotNull @Min(1) Integer quantity) {}
+public record StockRequestAndResponse(@NotBlank String name, @NotNull @Min(1) Long quantity) {}
